// src/command/sarif/sarif2.d.ts
export interface Sarif {
  /**
   * The URI of the JSON schema corresponding to the version.
   */
  $schema?: string;
  /**
   * The SARIF format version of this log file.
   */
  version: "2.1.0";
  /**
   * The set of runs contained in this log file.
   */
  runs: Run[];
  /**
   * References to external property files that share data between runs.
   */
  inlineExternalProperties?: ExternalProperties[];
  /**
   * Key/value pairs that provide additional information about the log file.
   */
  properties?: {
    /**
     * A set of distinct strings that provide additional information.
     */
    tags?: string[];
    [k: string]: unknown;
  };
}
// ...
export interface MultiformatMessageString {
  /**
   * A plain text message string or format string.
   */
  text: string;
  /**
   * A Markdown message string or format string.
   */
  markdown?: string;
  /**
   * Key/value pairs that provide additional information about the message.
   */
  properties?: {
    /**
     * A set of distinct strings that provide additional information.
     */
    tags?: string[];
    [k: string]: unknown;
  };
}
/**
 * Metadata that describes a specific report produced by the tool, as part of the analysis it provides or its runtime reporting.
 */
export interface ReportingDescriptor {
  /**
   * A stable, opaque identifier for the report.
   */
  id: string;
  /**
   * An array of stable, opaque identifiers by which this report was known in some previous version of the analysis tool.
   */
  deprecatedIds?: string[];
  /**
   * A unique identifier for the reporting descriptor in the form of a GUID.
   */
  guid?: string;
  /**
   * An array of unique identifies in the form of a GUID by which this report was known in some previous version of the analysis tool.
   */
  deprecatedGuids?: string[];
  /**
   * A report identifier that is understandable to an end user.
   */
  name?: string;
  /**
   * An array of readable identifiers by which this report was known in some previous version of the analysis tool.
   */
  deprecatedNames?: string[];
  /**
   * A message string or message format string rendered in multiple formats.
   */
  shortDescription?: {
    /**
     * A plain text message string or format string.
     */
    text: string;
    /**
     * A Markdown message string or format string.
     */
    markdown?: string;
    /**
     * Key/value pairs that provide additional information about the message.
     */
    properties?: {
      /**
       * A set of distinct strings that provide additional information.
       */
      tags?: string[];
      [k: string]: unknown;
    };
  };
  /**
   * A message string or message format string rendered in multiple formats.
   */
  fullDescription?: {
    /**
     * A plain text message string or format string.
     */
    text: string;
    /**
     * A Markdown message string or format string.
     */
    markdown?: string;
    /**
     * Key/value pairs that provide additional information about the message.
     */
    properties?: {
      /**
       * A set of distinct strings that provide additional information.
       */
      tags?: string[];
      [k: string]: unknown;
    };
  };
  /**
   * A set of name/value pairs with arbitrary names. Each value is a multiformatMessageString object, which holds message strings in plain text and (optionally) Markdown format. The strings can include placeholders, which can be used to construct a message in combination with an arbitrary number of additional string arguments.
   */
  messageStrings?: {
    [k: string]: MultiformatMessageString;
  };
  /**
   * Default reporting configuration information.
   */
  defaultConfiguration?: {
    /**
     * Specifies whether the report may be produced during the scan.
     */
    enabled?: boolean;
    /**
     * Specifies the failure level for the report.
     */
    level?: "none" | "note" | "warning" | "error";
    /**
     * Specifies the relative priority of the report. Used for analysis output only.
     */
    rank?: number;
    /**
     * Contains configuration information specific to a report.
     */
    parameters?: {
      /**
       * A set of distinct strings that provide additional information.
       */
      tags?: string[];
      [k: string]: unknown;
    };
    /**
     * Key/value pairs that provide additional information about the reporting configuration.
     */
    properties?: {
      /**
       * A set of distinct strings that provide additional information.
       */
      tags?: string[];
      [k: string]: unknown;
    };
  };
  /**
   * A URI where the primary documentation for the report can be found.
   */
  helpUri?: string;
  /**
   * A message string or message format string rendered in multiple formats.
   */
  help?: {
    /**
     * A plain text message string or format string.
     */
    text: string;
    /**
     * A Markdown message string or format string.
     */
    markdown?: string;
    /**
     * Key/value pairs that provide additional information about the message.
     */
    properties?: {
      /**
       * A set of distinct strings that provide additional information.
       */
      tags?: string[];
      [k: string]: unknown;
    };
  };
  /**
   * An array of objects that describe relationships between this reporting descriptor and others.
   */
  relationships?: ReportingDescriptorRelationship[];
  /**
   * Key/value pairs that provide additional information about the report.
   */
  properties?: {
    /**
     * A set of distinct strings that provide additional information.
     */
    tags?: string[];
    [k: string]: unknown;
  };
}
/**
 * Information about the relation of one reporting descriptor to another.
 */
export interface ReportingDescriptorRelationship {
  /**
   * A reference to the related reporting descriptor.
   */
  target:
    | {
        [k: string]: unknown;
      }
    | {
        [k: string]: unknown;
      }
    | {
        [k: string]: unknown;
      };
  /**
   * A set of distinct strings that categorize the relationship. Well-known kinds include 'canPrecede', 'canFollow', 'willPrecede', 'willFollow', 'superset', 'subset', 'equal', 'disjoint', 'relevant', and 'incomparable'.
   */
  kinds?: string[];
  /**
   * A description of the reporting descriptor relationship.
   */
  description?:
    | {
        [k: string]: unknown;
      }
    | {
        [k: string]: unknown;
      };
  /**
   * Key/value pairs that provide additional information about the reporting descriptor reference.
   */
  properties?: {
    /**
     * A set of distinct strings that provide additional information.
     */
    tags?: string[];
    [k: string]: unknown;
  };
}
// ...
export interface Artifact {
  /**
   * A short description of the artifact.
   */
  description?:
    | {
        [k: string]: unknown;
      }
    | {
        [k: string]: unknown;
      };
  /**
   * Specifies the location of an artifact.
   */
  location?: {
    /**
     * A string containing a valid relative or absolute URI.
     */
    uri?: string;
    /**
     * A string which indirectly specifies the absolute URI with respect to which a relative URI in the "uri" property is interpreted.
     */
    uriBaseId?: string;
    /**
     * The index within the run artifacts array of the artifact object associated with the artifact location.
     */
    index?: number;
    /**
     * A short description of the artifact location.
     */
    description?:
      | {
          [k: string]: unknown;
        }
      | {
          [k: string]: unknown;
        };
    /**
     * Key/value pairs that provide additional information about the artifact location.
     */
    properties?: {
      /**
       * A set of distinct strings that provide additional information.
       */
      tags?: string[];
      [k: string]: unknown;
    };
  };
  /**
   * Identifies the index of the immediate parent of the artifact, if this artifact is nested.
   */
  parentIndex?: number;
  /**
   * The offset in bytes of the artifact within its containing artifact.
   */
  offset?: number;
  /**
   * The length of the artifact in bytes.
   */
  length?: number;
  /**
   * The role or roles played by the artifact in the analysis.
   */
  roles?: (
    | "analysisTarget"
    | "attachment"
    | "responseFile"
    | "resultFile"
    | "standardStream"
    | "tracedFile"
    | "unmodified"
    | "modified"
    | "added"
    | "deleted"
    | "renamed"
    | "uncontrolled"
    | "driver"
    | "extension"
    | "translation"
    | "taxonomy"
    | "policy"
    | "referencedOnCommandLine"
    | "memoryContents"
    | "directory"
    | "userSpecifiedConfiguration"
    | "toolSpecifiedConfiguration"
    | "debugOutputFile"
  )[];
  /**
   * The MIME type (RFC 2045) of the artifact.
   */
  mimeType?: string;
  /**
   * The contents of the artifact.
   */
  contents?: {
    /**
     * UTF-8-encoded content from a text artifact.
     */
    text?: string;
    /**
     * MIME Base64-encoded content from a binary artifact, or from a text artifact in its original encoding.
     */
    binary?: string;
    /**
     * A message string or message format string rendered in multiple formats.
     */
    rendered?: {
      /**
       * A plain text message string or format string.
       */
      text: string;
      /**
       * A Markdown message string or format string.
       */
      markdown?: string;
      /**
       * Key/value pairs that provide additional information about the message.
       */
      properties?: {
        /**
         * A set of distinct strings that provide additional information.
         */
        tags?: string[];
        [k: string]: unknown;
      };
    };
    /**
     * Key/value pairs that provide additional information about the artifact content.
     */
    properties?: {
      /**
       * A set of distinct strings that provide additional information.
       */
      tags?: string[];
      [k: string]: unknown;
    };
  };
  /**
   * Specifies the encoding for an artifact object that refers to a text file.
   */
  encoding?: string;
  /**
   * Specifies the source language for any artifact object that refers to a text file that contains source code.
   */
  sourceLanguage?: string;
  /**
   * A dictionary, each of whose keys is the name of a hash function and each of whose values is the hashed value of the artifact produced by the specified hash function.
   */
  hashes?: {
    [k: string]: string;
  };
  /**
   * The Coordinated Universal Time (UTC) date and time at which the artifact was most recently modified. See "Date/time properties" in the SARIF spec for the required format.
   */
  lastModifiedTimeUtc?: string;
  /**
   * Key/value pairs that provide additional information about the artifact.
   */
  properties?: {
    /**
     * A set of distinct strings that provide additional information.
     */
    tags?: string[];
    [k: string]: unknown;
  };
}